put json files in separate folder within neo4j-test
