+++ neo4j-test/search-tests.js
@@ -9,11 +9,11 @@ import { neighbourhood, getInteractions, getNeighbouringNodes } from '../src/neo
 import { addDocumentToNeo4j } from '../src/neo4j/neo4j-document.js';
 import { deleteAllNodesAndEdges } from '../src/neo4j/test-functions.js';
 
-import goult1 from './doct_tests_1.json';
-import goult2 from './doct_tests_2.json';
-import goult3 from './doct_tests_3.json';
-import goult4 from './doct_tests_4.json';
-import goult5 from './doct_tests_5.json';
+import goult1 from './document/doct_tests_1.json';
+import goult2 from './document/doct_tests_2.json';
+import goult3 from './document/doct_tests_3.json';
+import goult4 from './document/doct_tests_4.json';
+import goult5 from './document/doct_tests_5.json';
 
 let rdbConn;
 let dbFix;
